Fix compatiblitiy with Openfire 4.8 - 5.0 (by dropping functionality)
This plugin has many issues raised against it, most of it related to features that:
- are incomplete (for example, this plugin reports on c2s activity on port 5222 but not 5223)
- are based on implementations that have long been replaced (e.g. MINA has been long gone, now replaced by Jetty)

To prevent carrying over an ever growing list of issues, in this commit, all incompatible measurements are simply dropped. This leaves the following statistics:

- Timestamp - Timestamp when the data was collected
- DB min - Minimum number of connections the pool may have
- DB max - Maximum number of connectiosn the pool may have
- DB current - Current number of connections the pool has
- DB used - Current number of connections being used
- Sessions - Current number of client sessions in the server

This is meager, but is more than not having any data at all (which was the status quo for the last few years.

fixes #14
fixes #12
fixes #11
fixes #10
fixes #9
fixes #8